moved all graph creation logic to it's own module
